Return JSON 500 for unhandled API errors
Express 5 forwards async handler rejections only to error middleware
registered after the route, so handleErrors registered before the routes
never saw them and unhandled errors fell through to the default HTML
handler. Split the Express app into app.ts (exported for supertest) with
index.ts as the serverless-express shim, and register handleErrors after
the routes.
